options.experimentalDynamicImport, move acorn plugin into separate file
